requirements.txt
Flask==3.0.3
python-telegram-bot==21.10
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
